split utils and errors
